Split base.rs
